+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/declarationIrBuilder.kt
@@ -1,0 +1,14 @@
+package com.javiersc.kotlin.compiler.extensions.ir
+
+import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
+import org.jetbrains.kotlin.backend.common.lower.DeclarationIrBuilder
+import org.jetbrains.kotlin.ir.declarations.IrSymbolOwner
+import org.jetbrains.kotlin.ir.symbols.IrSymbol
+
+public fun IrPluginContext.declarationIrBuilder(
+    symbol: IrSymbol,
+): DeclarationIrBuilder = DeclarationIrBuilder(this, symbol)
+
+public fun IrPluginContext.declarationIrBuilder(
+    element: IrSymbolOwner,
+): DeclarationIrBuilder = DeclarationIrBuilder(this, element.symbol)